+++ examples/nuxt/druxt-site/layouts/default.vue
@@ -1,3 +1,0 @@
-<template>
-  <DruxtSite theme="umami" />
-</template>