+++ packages/blocks/test/nuxtModule.test.js
@@ -4,12 +4,11 @@ jest.mock('../src/nuxtStorybook')
 
 const mock = {
   addModule: jest.fn(),
-  addPlugin: jest.fn(),
   nuxt: {
     hook: jest.fn((hook, fn) => {
       const arg = {
         'components:dirs': [],
-        'storybook:config': {}
+        'storybook:config': { stories: [] }
       }
       return fn(arg[hook])
     }),
@@ -18,11 +17,8 @@ const mock = {
 }
 
 test('Nuxt module', () => {
-  expect(() => { mock.DruxtBlocksNuxtModule() }).toThrow('Druxt settings missing.')
-
   mock.options = {
     druxt: {}
   }
   mock.DruxtBlocksNuxtModule()
-  expect(mock.addPlugin).toHaveBeenCalled()
 })