+++ packages/breadcrumb/test/nuxtModule.test.js
@@ -1,12 +1,11 @@
 import { DruxtBreadcrumbModule } from '../src/nuxtModule'
 
 const mock = {
-  addPlugin: jest.fn(),
   nuxt: {
     hook: jest.fn((hook, fn) => {
       const arg = {
         'components:dirs': [],
-        'storybook:config': {}
+        'storybook:config': { stories: [] }
       }
       return fn(arg[hook])
     }),
@@ -15,11 +14,8 @@ const mock = {
 }
 
 test('Nuxt module', () => {
-  expect(() => { mock.DruxtBreadcrumbModule() }).toThrow('Druxt settings missing.')
-
   mock.options = {
     druxt: {}
   }
   mock.DruxtBreadcrumbModule()
-  expect(mock.addPlugin).toHaveBeenCalled()
 })