+++ packages/entity/test/nuxtModule.test.js
@@ -4,12 +4,11 @@ jest.mock('../src/nuxtStorybook')
 
 const mock = {
   addModule: jest.fn(),
-  addPlugin: jest.fn(),
   nuxt: {
     hook: (hook, fn) => {
       const arg = {
         'components:dirs': [],
-        'storybook:config': {}
+        'storybook:config': { stories: [] }
       }
       return fn(arg[hook])
     }
@@ -18,12 +17,9 @@ const mock = {
 }
 
 test('Nuxt module', () => {
-  expect(() => { mock.DruxtEntityNuxtModule() }).toThrow('Druxt settings missing.')
-
   mock.options = {
     druxt: {},
     modules: [],
   }
   mock.DruxtEntityNuxtModule()
-  expect(mock.addPlugin).toHaveBeenCalled()
 })