+++ packages/menu/src/nuxtStorybook.js
@@ -28,9 +28,9 @@ export default async function ({ stories }) {
       const name = menu.attributes.drupal_internal__id
       const { description, label } = menu.attributes
 
-      const title = titleFn(['Druxt Menu', label])
+      const title = titleFn(['Druxt', 'Menu', label])
       addTemplate({
-        src: resolve(__dirname, '../nuxt/druxt-menu.stories.js'),
+        src: resolve(__dirname, '../templates/druxt-menu.stories.js'),
         fileName: `stories/druxt-menu.${name}.stories.js`,
         options: { description, label, name, title },
       })