+++ packages/schema/src/nuxtModule.js
@@ -25,29 +25,34 @@ import { DruxtSchema } from './schema'
  *   }
  * }
  *
- * @todo Add module level options.
  * @todo Document options.
  *
  * @param {object} moduleOptions - Nuxt.js module options object.
  */
-const DruxtSchemaNuxtModule = function () {
-  // Use root level Druxt options.
-  if (typeof this.options === 'undefined' || !this.options.druxt) {
-    throw new TypeError('Druxt settings missing.')
+const DruxtSchemaNuxtModule = function (moduleOptions = {}) {
+  // Set default options.
+  const options = {
+    baseUrl: moduleOptions.baseUrl,
+    ...(this.options || {}).druxt || {},
+    schema: {
+      ...((this.options || {}).druxt || {}).schema || {},
+      ...moduleOptions,
+    }
   }
-  const options = this.options.druxt
 
   // Add plugin.
   this.addPlugin({
-    src: resolve(__dirname, '../nuxt/plugin.js'),
+    src: resolve(__dirname, '../templates/plugin.js'),
     fileName: 'druxt-schema.js',
     options
   })
 
+  // Enable Vuex Store.
+  this.options.store = true
+
   // Add Vuex plugin.
-  // @TODO - Ensure Vuex store is available.
   this.addPlugin({
-    src: resolve(__dirname, '../nuxt/store.js'),
+    src: resolve(__dirname, '../templates/store.js'),
     fileName: 'store/druxt-schema.js',
     options
   })
@@ -62,7 +67,7 @@ const DruxtSchemaNuxtModule = function () {
       if (typeof schema === 'undefined') continue
 
       this.addTemplate({
-        src: resolve(__dirname, '../nuxt/schema.json'),
+        src: resolve(__dirname, '../templates/schema.json'),
         fileName: `schemas/${name}.json`,
         options: { schema }
       })