+++ packages/site/src/nuxtModule.js
@@ -1,4 +1,6 @@
-import { join } from 'path'
+import { existsSync } from 'fs'
+import { join, resolve } from 'path'
+import DruxtSiteStorybook from './nuxtStorybook'
 
 /**
  * Nuxt module function to install Druxt Site.
@@ -10,10 +12,20 @@ import { join } from 'path'
  *
  * @param {ModuleOptions} moduleOptions - The Nuxt.js module options.
  */
-const DruxtSiteNuxtModule = function () {
-  // Use root level Druxt options.
-  if (typeof this.options === 'undefined' || !this.options.druxt) {
-    throw new TypeError('Druxt settings missing.')
+const DruxtSiteNuxtModule = async function (moduleOptions = {}) {
+  // Set default options.
+  const options = {
+    baseUrl: moduleOptions.baseUrl,
+    ...(this.options || {}).druxt || {},
+    menu: {
+      jsonApiMenuItems: true,
+      ...((this.options || {}).druxt || {}).menu,
+    },
+    site: {
+      layout: true,
+      ...((this.options || {}).druxt || {}).site,
+      ...moduleOptions,
+    },
   }
 
   // Register components directories.
@@ -21,38 +33,40 @@ const DruxtSiteNuxtModule = function () {
     dirs.push({ path: join(__dirname, 'components') })
   })
 
-  // Add Nuxt.js modules.
-  const modules = [
-    '@nuxtjs/proxy',
-    'druxt',
+  // Add Druxt modules.
+  this.addModule(['druxt', options])
+  const druxtModules = [
     'druxt-blocks',
     'druxt-breadcrumb',
     'druxt-entity',
     'druxt-menu',
-    'druxt-router',
+    'druxt-router/nuxt',
     'druxt-schema',
     'druxt-views'
   ]
-  for (const key in modules) {
-    this.addModule(modules[key])
+  for (const module of druxtModules) {
+    this.addModule([module, { baseUrl: options.baseUrl, ...options[module.split('-')[1]] || {}}])
   }
 
   // Setup proxy for 'sites/default/files'.
-  // @todo {@link https://github.com/nuxt-community/proxy-module/issues/80|Suppress warning}
   if (typeof this.options.proxy === 'undefined') {
-    this.options.proxy = [this.options.druxt.baseUrl + '/sites/default/files']
+    this.options.proxy = [options.baseUrl + '/sites/default/files']
   }
-
-  // Enable JSON:API Menu items by default.
-  if (typeof ((this.options.druxt || {}).menu || {}).jsonApiMenuItems === 'undefined') {
-    this.options.druxt.menu = {
-      ...this.options.druxt.menu,
-      ...{ jsonApiMenuItems: true }
-    }
-  }
+  this.addModule('@nuxtjs/proxy')
 
   // Enable Vuex Store.
   this.options.store = true
+
+  // Add default layout.
+  if (!(await existsSync(resolve(this.options.srcDir, this.options.dir.layouts))) && options.site.layout) {
+    this.addLayout(resolve(__dirname, './layouts/default.vue'), 'default')
+  }
+
+  // Nuxt Storybook.
+  this.nuxt.hook('storybook:config', async ({ stories }) => {
+    stories.push('druxt-site/dist/components/*.stories.mjs')
+    await DruxtSiteStorybook.call(this, { options, stories })
+  })
 }
 
 DruxtSiteNuxtModule.meta = require('../package.json')